+++ test/Armada.Test.Unit/Program.cs
@@ -338,6 +338,7 @@ namespace Armada.Test.Unit
             runner.AddSuite(new TypedTestCoversAdapterTests());
             runner.AddSuite(new TypedLintFindingAdapterTests());
             runner.AddSuite(new PriorArtRetrieverTests());
+            runner.AddSuite(new GitPriorArtSourceTests());
             runner.AddSuite(new TypedPriorArtAdapterTests());
             runner.AddSuite(new FileMemoryCandidateProposalWriterTests());
             runner.AddSuite(new OpenCodeServerInferenceClientTests());